+++ test/core/core-test-util.ts
@@ -82,13 +82,12 @@ class BogusService implements TService {
   }
 }
 
-export const commonArgs: Omit<CoreArgs, "oldTarget" | "src" | "srcCache"> = {
+export const commonArgs: Omit<CoreArgs, "oldTarget" | "src"> = {
   service: bogusTranslateName as TServiceType,
   serviceConfig: "invalid-core-key",
   matcher: "icu",
   srcLng: "en",
   targetLng: "de",
-  overwriteOutdated: true,
 };
 
 export async function translateCoreAssert(
@@ -127,8 +126,4 @@ export function generateSubTSets(fullTSet: TSet): TSet[] {
   }
   expect(subTSets.length).toBe(Math.pow(2, fullSet.length));
   return subTSets;
-}
-
-export function getRandomBoolean(): boolean {
-  return Math.random() >= 0.5;
 }