+++ test/core/up-to-date-cache.test.ts
@@ -7,11 +7,10 @@ import {
 import { CoreArgs, CoreResults } from "../../src/core/core-definitions";
 import { toStrictEqualMapOrder } from "../test-util/to-strict-equal-map-order";
 
-test("up-to-date cache, no target", async () => {
+test("no target", async () => {
   const args: CoreArgs = {
     ...commonArgs,
     src: enSrc,
-    srcCache: enSrc,
     oldTarget: null,
   };
   const expectRes: CoreResults = {
@@ -22,7 +21,6 @@ test("up-to-date cache, no target", async () => {
       deleted: null,
     },
     newTarget: deTarget,
-    newSrcCache: args.src,
     serviceInvocation: {
       inputs: enSrc,
       results: deTarget,
@@ -36,12 +34,10 @@ test("up-to-date cache, up-to-date target", async () => {
   const args: CoreArgs = {
     ...commonArgs,
     src: enSrc,
-    srcCache: enSrc,
     oldTarget: deTarget,
   };
   const expectRes: CoreResults = {
     newTarget: deTarget,
-    newSrcCache: args.src,
     changeSet: {
       added: new Map(),
       updated: new Map(),
@@ -53,35 +49,3 @@ test("up-to-date cache, up-to-date target", async () => {
   const res = await translateCoreAssert(args);
   toStrictEqualMapOrder(res, expectRes);
 });
-
-// const staleTarget: TSet = new Map([
-//   ["1", "fgebg"],
-//   ["2", "Wdbhdelt"],
-//   ["3", "fwfsfs"],
-//   ["4", "stsd"],
-//   ["5", "sfsef"],
-//   ["6", "rrw"],
-//   ["leftover", "Outdated"],
-// ]);
-
-// test("up-to-date cache, do not delete stale", async () => {
-//   const args: CoreArgs = {
-//     ...commonArgs,
-//     src: enSrc,
-//     srcCache: enSrc,
-//     oldTarget: staleTarget,
-//   };
-//   const expectRes: CoreResults = {
-//     newTarget: staleTarget,
-//     newSrcCache: args.src,
-//     changeSet: {
-//       added: new Map(),
-//       updated: new Map(),
-//       skipped: new Map(),
-//       deleted: null,
-//     },
-//     serviceInvocation: null,
-//   };
-//   const res = await translateCoreAssert(args);
-//   toStrictEqualMapOrder(res, expectRes);
-// });